Data.Avro.Deriving.NormSchema

Documentation

extractDerivables :: Schema -> [Schema]

Extracts all the records from the schema (flattens the schema) Named types get resolved when needed to include at least one "inlined" schema in each record and to make each record self-contained. Note: Namespaces are not really supported in this version. All the namespaces (including inlined into full names) will be ignored during names resolution.

getTypes :: Schema -> [(TypeName, Schema)]

normSchema :: Schema -> State (Map TypeName Schema) Schema