Data.Avro.EitherN

Documentation

data Either3 a b c

Constructors

E3_1 a

E3_2 b

E3_3 c

data Either4 a b c d

Constructors

E4_1 a

E4_2 b

E4_3 c

E4_4 d

data Either5 a b c d e

Constructors

E5_1 a

E5_2 b

E5_3 c

E5_4 d

E5_5 e

data Either6 a b c d e f

Constructors

E6_1 a

E6_2 b

E6_3 c

E6_4 d

E6_5 e

E6_6 f

data Either7 a b c d e f g

Constructors

E7_1 a

E7_2 b

E7_3 c

E7_4 d

E7_5 e

E7_6 f

E7_7 g

data Either8 a b c d e f g h

Constructors

E8_1 a

E8_2 b

E8_3 c

E8_4 d

E8_5 e

E8_6 f

E8_7 g

E8_8 h

data Either9 a b c d e f g h i

Constructors

E9_1 a

E9_2 b

E9_3 c

E9_4 d

E9_5 e

E9_6 f

E9_7 g

E9_8 h

E9_9 i

data Either10 a b c d e f g h i j

Constructors

E10_1 a

E10_2 b

E10_3 c

E10_4 d

E10_5 e

E10_6 f

E10_7 g

E10_8 h

E10_9 i

E10_10 j

putIndexedValue :: ToAvro a => Int -> Vector Schema -> a -> Builder