Data.Avro.Encoding.FromAvro

Documentation

class FromAvro a where

Descrives how to convert a given intermediate Value into a Haskell data type.

Methods

fromAvro :: Value -> Either String a

For internal use

data Value

An intermediate data structute for decoding between Avro bytes and Haskell types.

Because reader and writer schemas, and therefore expected data types and layout can be different, deserialising bytes into Haskell types directly is not possible.

To overcome this issue this intermediate data structure is used: bytes are decoded into values of type Value (using reader's layout and rules) and then translated to target Haskell types using FromAvro type class machinery.

Constructors

Null

Boolean Bool

Int ReadSchema Int32

Long ReadSchema Int64

Float ReadSchema Float

Double ReadSchema Double

Bytes ReadSchema ByteString

String ReadSchema Text

Array (Vector Value)

Map (HashMap Text Value)

Record ReadSchema (Vector Value)

Union ReadSchema Int Value

Fixed ReadSchema ByteString

Enum ReadSchema Int Text

getValue :: ReadSchema -> Get Value

describeValue :: Value -> String

Descrive the value in a way that is safe to use in error messages (i.e. do not print values)