Data.Avro.Internal.Get

Description

An internal module that contains common decoding functionality that is shared between Lazy and Strict decoders, as well as generic Get monad helpers.

Documentation

getBoolean :: Get Bool

getInt :: Get Int32

Get a 32-bit int (zigzag encoded, max of 5 bytes)

getLong :: Get Int64

Get a 64 bit int (zigzag encoded, max of 10 bytes)

getZigZag :: (Bits i, Integral i, DecodeRaw i) => Get i

Get an zigzag encoded integral value consuming bytes till the msb is 0.

getBytes :: Get ByteString

getBytesLazy :: Get ByteString

getString :: Get Text

getFloat :: Get Float

getDouble :: Get Double

decodeBlocks :: Get a -> Get [a]

Avro encodes arrays and maps as a series of blocks. Each block starts with a count of the elements in the block. A series of blocks is always terminated with an empty block (encoded as a 0).

sFromIntegral :: forall a b m. (Monad m, Bounded a, Bounded b, Integral a, Integral b) => a -> m b