Data.Avro.Schema.Deconflict

Documentation

deconflict :: Schema -> Schema -> Either String ReadSchema

deconflict writer reader will produce a schema that can decode with the writer's schema into the form specified by the reader's schema.

Schema resolution rules are described by the specification: https://avro.apache.org/docs/current/spec.html#Schema+Resolution