Data.Avro.Schema.ReadSchema

Documentation

data ReadSchema

This type represents a deconflicted version of a Schema. Schema resolution is described in Avro specification: https://avro.apache.org/docs/current/spec.html#Schema+Resolution

This library represents "normal" schema and "deconflicted" schema as different types to avoid confusion between these two usecases (we shouldn't serialise values with such schema) and to be able to accomodate some extra information that links between how data is supposed transformed between what reader wants and what writer has.

Constructors

Null

Boolean

Int

Fields

logicalTypeI :: Maybe LogicalTypeInt

Long

Fields

longReadFrom :: ReadLong

logicalTypeL :: Maybe LogicalTypeLong

Float

Fields

floatReadFrom :: ReadFloat

Double

Fields

doubleReadFrom :: ReadDouble

Bytes

Fields

logicalTypeB :: Maybe LogicalTypeBytes

String

Fields

logicalTypeS :: Maybe LogicalTypeString

Array

Fields

item :: ReadSchema

Map

Fields

values :: ReadSchema

NamedType TypeName

Record

Fields

name :: TypeName

aliases :: [TypeName]

doc :: Maybe Text

fields :: [ReadField]

Enum

Fields

name :: TypeName

aliases :: [TypeName]

doc :: Maybe Text

symbols :: Vector Text

Union

Fields

options :: Vector (Int, ReadSchema)

Order of values represents order in the writer schema, an index represents order in a reader schema

Fixed

Fields

name :: TypeName

aliases :: [TypeName]

size :: Int

logicalTypeF :: Maybe LogicalTypeFixed

FreeUnion

Fields

pos :: Int

ty :: ReadSchema

data ReadField

Deconflicted record field.

Constructors

ReadField

Fields

fldName :: Text

fldAliases :: [Text]

fldDoc :: Maybe Text

fldOrder :: Maybe Order

fldStatus :: FieldStatus

How the value of this field should be treated. See FieldStatus documentation.

fldType :: ReadSchema

fldDefault :: Maybe DefaultValue

data ReadLong

How to decode a value of target type Long. This type controls how many bits are needed to be read from the encoded bytestring. The number of bits can be different depending on differences between reader and writer schemas.

The rules are described in https://avro.apache.org/docs/current/spec.html#Schema+Resolution

Constructors

LongFromInt
Read Int (32 bits) and cast it to Long (Rule: int is promotable to long, float, or double)

ReadLong
Read Long (64 bits) and use as is

data ReadFloat

How to decode a value of target type Float. This type controls how many bits are needed to be read from the encoded bytestring. The number of bits can be different depending on differences between reader and writer schemas.

The rules are described in https://avro.apache.org/docs/current/spec.html#Schema+Resolution

Constructors

FloatFromInt
Read Int (32 bits) and cast it to Float

FloatFromLong
Read Long (64 bits) and cast it to Float (Rule: long is promotable to float or double)

ReadFloat
Read Float and use as is

data ReadDouble

How to decode a value of target type Double. This type controls how many bits are needed to be read from the encoded bytestring. The number of bits can be different depending on differences between reader and writer schemas.

The rules are described in https://avro.apache.org/docs/current/spec.html#Schema+Resolution

Constructors

DoubleFromInt
Read Int (32 bits) and cast it to Double (Rule: int is promotable to long, float, or double)

DoubleFromFloat
Read Float (64 bits) and cast it to Double (Rule: float is promotable to float or double)

DoubleFromLong
Read Long (64 bits) and cast it to Double (Rule: long is promotable to float or double)

ReadDouble

fromSchema :: Schema -> ReadSchema

Converts Avro Schema to ReaderSchema trivially. This function is useful when no deconflicting is required.

fromField :: FieldStatus -> Field -> ReadField

extractBindings :: ReadSchema -> HashMap TypeName ReadSchema

extractBindings schema traverses a schema and builds a map of all declared types.

Types declared implicitly in record field definitions are also included. No distinction is made between aliases and normal names.

data Decimal

The decimal logical type represents arbitrary-precision decimal numbers. Numbers are represented as unscaled * (10 ** -scale) where scale is part of the logical type and unscaled is an integer represented by the underlying primitive type.

Instances of the decimal logical type need to specify a scale and precision.

decimal can be encoded as one of several different primitive types:

bytes

fixed

long

int

For long and int, unscaled is the underlying number.

For bytes and fixed, unscaled is represented as a two's-complement signed integer in big-endian byte order.

Note: int and long representations for decimal are not part of the current Avro specification, but they are supported by some language implementations including the official Java library. Implementations that do not support this should ignore the logical type and use the underlying primitive type instead.

Constructors

Decimal

Fields

precision :: Integer

The maximum number of digits that can be represented by this decimal type.

precision > 0

scale :: Integer

The scale in unscaled * (10 ** -scale) for this type.

0 ≤ scale ≤ precision

newtype LogicalTypeBytes

Constructors

DecimalB Decimal
An arbitrary-precision signed decimal number. See Decimal.

data LogicalTypeFixed

Constructors

DecimalF Decimal
An arbitrary-precision signed decimal number. See Decimal.

Duration
An interval of time, represented as some number of months, days and milliseconds.

Encoded as three little-endian unsigned integers for months, days and milliseconds respectively.

data LogicalTypeInt

Constructors

DecimalI Decimal
An arbitrary-precision signed decimal number. See Decimal.

Date
A date (eg 2020-01-10) with no timezone/locale.

Encoded as the number of days before/after the Unix epoch (1970-01-01).

TimeMillis
A time of day with millisecond precision.

Encoded as the number of milliseconds after midnight.

data LogicalTypeLong

Constructors

DecimalL Decimal
An arbitrary-precision signed decimal number. See Decimal.

TimeMicros
A time of day with microsecond precision.

Encoded as the number of microseconds after midnight.

TimestampMillis
A UTC timestamp with millisecond precision.

Encoded as the number of milliseconds before/after the Unix epoch (1970-01-01 00:00:00.000).

TimestampMicros
A UTC timestamp with microsecond precision.

Encoded as the number of microseconds before/after the Unix epoch (1970-01-01 00:00:00.000000).

LocalTimestampMillis
A timestamp in the local timezone, whatever that happens to be, with millisecond precision.

Encoded as the number of milliseconds before/after the Unix epoch (1970-01-01 00:00:00.000).

LocalTimestampMicros
A timestamp in the local timezone, whatever that happens to be, with microsecond precision.

Encoded as the number of microseconds before/after the Unix epoch (1970-01-01 00:00:00.000000).

data LogicalTypeString

Constructors

UUID
A Universally Unique Identifier (UUID).

Encoded as a string that is valid according to RFC 4122.

data FieldStatus

Depending on differences between reader and writer schemas, a record field can be found:

Present in the reader schema but missing from the writer schema. In this case the reader field is marked as Defaulted with the default value from the reader schema. An index value represents the position of the field in the reader schema.

Present in the writer schema but missing from the reader schema. In this case the record field is marked as Ignored: the corresponding bytes still need to be read from the payload (to advance the position in a bytestring), but the result is discarded.

Present in both reader and writer schemas. In this case the field is marked to be read AsIs with an index that represents the field's position in the reader schema.

Constructors

AsIs Int

Ignored

Defaulted Int DefaultValue