Data.Avro.Schema.Schema

Description

Avro Schemas, represented here as values of type Schema, describe the serialization and de-serialization of values.

In Avro schemas are compose-able such that encoding data under a schema and decoding with a variant, such as newer or older version of the original schema, can be accomplished by using the Deconflict module.

Schema description types

data Schema

N.B. It is possible to create a Haskell value (of Schema type) that is not a valid Avro schema by violating one of the above or one of the conditions called out in validateSchema.

Constructors

Null

Boolean

Int

Fields

logicalTypeI :: Maybe LogicalTypeInt

Long

Fields

logicalTypeL :: Maybe LogicalTypeLong

Float

Double

Bytes

Fields

logicalTypeB :: Maybe LogicalTypeBytes

String

Fields

logicalTypeS :: Maybe LogicalTypeString

Array

Fields

item :: Schema

Map

Fields

values :: Schema

NamedType TypeName

Record

Fields

name :: TypeName

aliases :: [TypeName]

doc :: Maybe Text

fields :: [Field]

Enum

Fields

name :: TypeName

aliases :: [TypeName]

doc :: Maybe Text

symbols :: Vector Text

Union

Fields

options :: Vector Schema

Fixed

Fields

name :: TypeName

aliases :: [TypeName]

size :: Int

logicalTypeF :: Maybe LogicalTypeFixed

Bundled Patterns

pattern Int' :: Schema

pattern Long' :: Schema

pattern Bytes' :: Schema

pattern String' :: Schema

data DefaultValue

Constructors

DNull

DBoolean !Bool

DInt Schema Int32

DLong Schema Int64

DFloat Schema Float

DDouble Schema Double

DBytes Schema ByteString

DString Schema Text

DArray (Vector DefaultValue)
Dynamically enforced monomorphic type.

DMap (HashMap Text DefaultValue)
Dynamically enforced monomorphic type

DRecord Schema (HashMap Text DefaultValue)

DUnion (Vector Schema) Schema DefaultValue
Set of union options, schema for selected option, and the actual value.

DFixed Schema !ByteString

DEnum Schema Int Text
An enum is a set of the possible symbols (the schema) and the selected symbol

data Field

Constructors

Field

Fields

fldName :: Text

fldAliases :: [Text]

fldDoc :: Maybe Text

fldOrder :: Maybe Order

fldType :: Schema

fldDefault :: Maybe DefaultValue

data Order

Constructors

Ascending

Descending

Ignore

data TypeName

A named type in Avro has a name and, optionally, a namespace.

A name is a string that starts with an ASCII letter or underscore followed by letters, underscores and digits:

name ::= [A-Za-z_][A-Za-z0-9_]*

Examples include "_foo7", Bar_ and "x".

A namespace is a sequence of names with the same lexical structure. When written as a string, the components of a namespace are separated with dots ("com.example").

TypeName represents a fullname—a name combined with a namespace. These are written and parsed as dot-separated strings. The TypeName TN Foo ["com", "example"] is rendered as "com.example.Foo".

Fullnames have to be globally unique inside an Avro schema.

A namespace of [] or [""] is the "null namespace". In avro an explicitly null-namespaced identifier is written as ".Foo"

Constructors

TN

Fields

baseName :: Text

namespace :: [Text]

data Decimal

The decimal logical type represents arbitrary-precision decimal numbers. Numbers are represented as unscaled * (10 ** -scale) where scale is part of the logical type and unscaled is an integer represented by the underlying primitive type.

Instances of the decimal logical type need to specify a scale and precision.

decimal can be encoded as one of several different primitive types:

bytes

fixed

long

int

For long and int, unscaled is the underlying number.

For bytes and fixed, unscaled is represented as a two's-complement signed integer in big-endian byte order.

Note: int and long representations for decimal are not part of the current Avro specification, but they are supported by some language implementations including the official Java library. Implementations that do not support this should ignore the logical type and use the underlying primitive type instead.

Constructors

Decimal

Fields

precision :: Integer

The maximum number of digits that can be represented by this decimal type.

precision > 0

scale :: Integer

The scale in unscaled * (10 ** -scale) for this type.

0 ≤ scale ≤ precision

newtype LogicalTypeBytes

Constructors

DecimalB Decimal
An arbitrary-precision signed decimal number. See Decimal.

data LogicalTypeFixed

Constructors

DecimalF Decimal
An arbitrary-precision signed decimal number. See Decimal.

Duration
An interval of time, represented as some number of months, days and milliseconds.

Encoded as three little-endian unsigned integers for months, days and milliseconds respectively.

data LogicalTypeInt

Constructors

DecimalI Decimal
An arbitrary-precision signed decimal number. See Decimal.

Date
A date (eg 2020-01-10) with no timezone/locale.

Encoded as the number of days before/after the Unix epoch (1970-01-01).

TimeMillis
A time of day with millisecond precision.

Encoded as the number of milliseconds after midnight.

data LogicalTypeLong

Constructors

DecimalL Decimal
An arbitrary-precision signed decimal number. See Decimal.

TimeMicros
A time of day with microsecond precision.

Encoded as the number of microseconds after midnight.

TimestampMillis
A UTC timestamp with millisecond precision.

Encoded as the number of milliseconds before/after the Unix epoch (1970-01-01 00:00:00.000).

TimestampMicros
A UTC timestamp with microsecond precision.

Encoded as the number of microseconds before/after the Unix epoch (1970-01-01 00:00:00.000000).

LocalTimestampMillis
A timestamp in the local timezone, whatever that happens to be, with millisecond precision.

Encoded as the number of milliseconds before/after the Unix epoch (1970-01-01 00:00:00.000).

LocalTimestampMicros
A timestamp in the local timezone, whatever that happens to be, with microsecond precision.

Encoded as the number of microseconds before/after the Unix epoch (1970-01-01 00:00:00.000000).

data LogicalTypeString

Constructors

UUID
A Universally Unique Identifier (UUID).

Encoded as a string that is valid according to RFC 4122.

renderFullname :: TypeName -> Text

Render a fullname as a dot separated string.

> renderFullname (TN Foo ["com", "example"])
"com.example.Foo"

> renderFullname (TN Foo [])
".Foo"

parseFullname :: Text -> TypeName

Parses a fullname into a TypeName, assuming the string representation is valid.

> parseFullname "com.example.Foo"
TN { baseName = Foo, components = ["com", "example"] }

mkEnum

Arguments

:: TypeName
The name of the enum (includes namespace).

-> [TypeName]
Aliases for the enum (if any).

-> Maybe Text
Optional documentation for the enum.

-> [Text]
The symbols of the enum.

-> Schema

Build an Enum value from its components.

mkUnion :: NonEmpty Schema -> Schema

mkUnion subTypes Defines a union of the provided subTypes. N.B. it is invalid Avro to include another union or to have more than one of the same type as a direct member of the union. No check is done for this condition!

validateSchema :: Schema -> Parser ()

Placeholder NO-OP function!

Validates a schema to ensure:

All types are defined

Unions do not directly contain other unions

Unions are not ambiguous (may not contain more than one schema with the same type except for named types of record, fixed and enum)

Default values for unions can be cast as the type indicated by the first structure.

Default values can be cast/de-serialize correctly.

Named types are resolvable

Lower level utilities

typeName :: Schema -> Text

Get the name of the type. In the case of unions, get the name of the first value in the union schema.

typeAliases :: Schema -> [TypeName]

Get the aliases of the type.

buildTypeEnvironment

Arguments

:: Applicative m

=> (TypeName -> m Schema)
Callback to handle type names not in the schema.

-> Schema
The schema that we're generating a lookup function for.

-> TypeName

-> m Schema

buildTypeEnvironment schema builds a function mapping type names to the types declared in the traversed schema.

This mapping includes both the base type names and any aliases they have. Aliases and normal names are not differentiated in any way.

extractBindings :: Schema -> HashMap TypeName Schema

extractBindings schema traverses a schema and builds a map of all declared types.

Types declared implicitly in record field definitions are also included. No distinction is made between aliases and normal names.

data Result a

Constructors

Success a

Error String

badValue :: Show t => t -> String -> Result a

resultToEither :: Result b -> Either String b

matches :: Schema -> Schema -> Bool

Checks that two schemas match. This is like equality of schemas, except NamedTypes match against other types with the same name.

This extends recursively: two records match if they have the same name, the same number of fields and the fields all match.

parseBytes :: Text -> Result ByteString

Parses a string literal into a bytestring in the format expected for bytes and fixed values. Will fail if every character does not have a codepoint between 0 and 255.

serializeBytes :: ByteString -> Text

Turn a ByteString into a Text that matches the format Avro expects from bytes and fixed literals in JSON. Each byte is mapped to a single Unicode codepoint between 0 and 255.

parseAvroJSON

Arguments

:: (Schema -> Value -> Result DefaultValue)
How to handle unions. The way unions are formatted in JSON depends on whether we're parsing a normal Avro object or we're parsing a default declaration in a schema.

This function will only ever be passed Union schemas. It should error out if this is not the case—it represents a bug in this code.

-> (TypeName -> Maybe Schema)

-> Schema

-> Value

-> Result DefaultValue

Parse JSON-encoded avro data.

overlay :: Schema -> Schema -> Schema

Merge two schemas to produce a third. Specifically, overlay schema reference fills in NamedTypes in schema using any matching definitions from reference.

subdefinition :: Schema -> Text -> Maybe Schema

Extract the named inner type definition as its own schema.

expandNamedTypes :: Schema -> Schema